Updated the json dict to follow the EU samples.
We have changed some of the keys of the dict that were in English. Namely, we have have change the headers of the top investments and industry tables. We will modify these headers on Alddin Explore so they are displayed in English properly but then cause no error when translating them.